+++ clang/lib/3C/CastPlacement.cpp
@@ -13,7 +13,7 @@
 #include "clang/3C/3CGlobalOptions.h"
 #include "clang/3C/ConstraintResolver.h"
 #include "clang/3C/Utils.h"
-#include <clang/Tooling/Refactoring/SourceCode.h>
+#include <clang/Tooling/Transformer/SourceCode.h>
 
 using namespace clang;
